requirements.txt
boto3>=1.28.0
requests>=2.31.0
lxml>=4.9.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0
pandas>=2.0.0
jupyter>=1.0.0
rouge-score>=0.1.2
fastapi>=0.100.0
uvicorn>=0.23.0
pyyaml>=6.0
